backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
pymongo>=4.3
python-dotenv>=1.0.0
dnspython>=2.0.0
pymongo>=4.3
mongoengine>=0.27.0
